+++ airbyte-cdk/python/airbyte_cdk/sources/utils/transform.py
@@ -9,7 +9,8 @@ from typing import Any, Callable, Dict, Mapping, Optional
 
 from jsonschema import Draft7Validator, ValidationError, validators
 
-json_to_python = {"string": str, "number": float, "integer": int, "boolean": bool, "null": type(None), "object": dict, "array": list}
+json_to_python_simple = {"string": str, "number": float, "integer": int, "boolean": bool, "null": type(None)}
+json_to_python = json_to_python_simple | {"object": dict, "array": list}
 python_to_json = {v: k for k, v in json_to_python.items()}
 
 logger = logging.getLogger("airbyte")
@@ -117,6 +118,10 @@ class TypeTransformer:
                 if isinstance(original_item, str):
                     return strtobool(original_item) == 1
                 return bool(original_item)
+            elif target_type == "array":
+                item_types = set(subschema.get("items", {}).get("type", set()))
+                if item_types.issubset(json_to_python_simple) and type(original_item) in json_to_python_simple.values():
+                    return [original_item]
         except (ValueError, TypeError):
             return original_item
         return original_item