+++ airbyte-cdk/python/CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## 0.1.86
+- TypeTransformer now converts simple types to array of simple types
 - TypeTransformer make warning message more informative
 
 ## 0.1.85